move files to their own subdirectory
Currently, each source file is a monolithic, single collection of all types
and functions used in the project. This is ugly, and hurts code reusability
to some degree.

The plan is to refactor these into several header and source files, and to
provide a makefile for each project. This commit is a preparation step for
that.